# src/main/resources/openapi/catalog-api.yaml
openapi: 3.0.3
info:
  title: 5Vision Catalog API
  version: 1.6.0
  description: Advanced contract for nature cards, taxonomies, and user favorites.
security:
  - bearerAuth: []
tags:
  - name: Cards
    description: Card management and retrieval
  - name: Favorites
    description: User favorite operations
  - name: Taxonomy
    description: Category and tag administration

paths:
  /api/v1/cards:
    get:
      tags: [Cards]
      summary: List and filter published cards (public)
      operationId: getCards
      security: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - name: sort
          in: query
          description: "Format: property,asc|desc. Allowed: createdAt, updatedAt, title, viewCount, favoriteCount"
          schema: { type: string, default: "createdAt,desc" }
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: tagId
          in: query
          schema: { type: string, format: uuid }
        - name: isPremium
          in: query
          schema: { type: boolean }
        - name: authorId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Paginated cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Cards]
      summary: Create card (Author/Admin)
      operationId: createCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/cards/my-cards:
    get:
      tags: [Cards]
      summary: List cards owned by the authenticated user
      operationId: getMyCards
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - name: status
          in: query
          schema: { type: string, enum: [DRAFT, PUBLISHED, ARCHIVED] }
        - name: sort
          in: query
          schema: { type: string, default: "updatedAt,desc" }
      responses:
        '200':
          description: Paginated cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/cards/favorites:
    get:
      tags: [Favorites]
      summary: List user's favorited cards
      operationId: getFavoriteCards
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: Paginated favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # -------------------------------------------------------------------
  # Bulk delete. Same shape as the media bulk-delete so the frontend can
  # reuse the same UI and the same service helper signature.
  # Placed before /{id} for clarity; Spring's matcher picks literal
  # segments over path variables anyway.
  # -------------------------------------------------------------------
  /api/v1/cards/bulk-delete:
    post:
      tags: [Cards]
      summary: Delete multiple cards in a single request (Owner/Admin)
      operationId: bulkDeleteCards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteCardsRequest'
      responses:
        '200':
          description: Per-item delete result. Partial failures are reported in `errors`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDeleteCardsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/cards/{id}:
    get:
      tags: [Cards]
      summary: Get card by ID (public)
      operationId: getCardById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Card details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Cards]
      summary: Update card (Owner/Admin)
      operationId: updateCard
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Cards]
      summary: Delete card (Owner/Admin)
      operationId: deleteCard
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/cards/{id}/view:
    post:
      tags: [Cards]
      summary: Increment view count for a card (public)
      operationId: incrementCardView
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: View count incremented
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/cards/{id}/favorite:
    post:
      tags: [Favorites]
      summary: Favorite a card
      operationId: favoriteCard
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Favorited
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Favorites]
      summary: Unfavorite a card
      operationId: unfavoriteCard
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Unfavorited
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/categories:
    get:
      tags: [Taxonomy]
      summary: List all categories (public)
      operationId: getCategories
      security: []
      parameters:
        - name: tree
          in: query
          description: If true, returns only top-level categories with their children nested.
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Taxonomy]
      summary: Create a new category (Admin)
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/categories/{id}:
    put:
      tags: [Taxonomy]
      summary: Update category (Admin)
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Taxonomy]
      summary: Delete category (Admin)
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Category deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/tags:
    get:
      tags: [Taxonomy]
      summary: List tags (public)
      operationId: getTags
      security: []
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Taxonomy]
      summary: Create tag (Admin)
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/tags/{id}:
    put:
      tags: [Taxonomy]
      summary: Update tag (Admin)
      operationId: updateTag
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '200':
          description: Tag updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Taxonomy]
      summary: Delete tag (Admin)
      operationId: deleteTag
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Tag deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued JWT token

  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    PagedCardResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardSummaryResponse'
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        isLast:
          type: boolean

    CreateCardRequest:
      type: object
      required: [title, speciesScientificName, primaryMediaId]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 5000
        speciesCommonName:
          type: string
          maxLength: 255
        speciesScientificName:
          type: string
          minLength: 1
          maxLength: 255
        conservationStatus:
          type: string
          maxLength: 50
        locationName:
          type: string
          maxLength: 255
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        primaryMediaId:
          type: string
          format: uuid
        thumbnailMediaId:
          type: string
          format: uuid
        isPremium:
          type: boolean
          default: false
        price:
          type: number
          format: double
          minimum: 0
        categoryIds:
          type: array
          items:
            type: string
            format: uuid
          uniqueItems: true
        tagIds:
          type: array
          items:
            type: string
            format: uuid
          uniqueItems: true

    UpdateCardRequest:
      allOf:
        - $ref: '#/components/schemas/CreateCardRequest'
        - type: object
          properties:
            status:
              type: string
              enum: [DRAFT, PUBLISHED, ARCHIVED]

    AuthorSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }

    MediaSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        cdnUrl: { type: string }
        thumbnailUrl: { type: string }
        largeUrl: { type: string }
        type: { type: string, enum: [ PHOTO, VIDEO, THUMBNAIL ] }

    CardSummaryResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        speciesCommonName: { type: string }
        speciesScientificName: { type: string }
        locationName: { type: string }
        isPremium: { type: boolean }
        price: { type: number, format: double }
        viewCount: { type: integer, format: int64 }
        favoriteCount: { type: integer, format: int64 }
        # The card's PRIMARY media type. This is the authoritative signal a
        # client should use to decide whether a card is a photo or a video —
        # the thumbnail is often a plain PHOTO poster for a VIDEO card, so
        # thumbnailMedia.type alone is not reliable.
        primaryMediaType:
          type: string
          enum: [PHOTO, VIDEO, THUMBNAIL]
        # Duration of the primary media in seconds, present only for videos.
        # Lets clients render the duration badge before the mp4 metadata loads.
        durationSeconds:
          type: integer
          minimum: 0
        thumbnailMedia: { $ref: '#/components/schemas/MediaSummary' }
        createdAt: { type: string, format: date-time }

    CardDetailResponse:
      allOf:
        - $ref: '#/components/schemas/CardSummaryResponse'
        - type: object
          properties:
            description: { type: string }
            conservationStatus: { type: string }
            latitude: { type: number, format: double }
            longitude: { type: number, format: double }
            status: { type: string, enum: [DRAFT, PUBLISHED, ARCHIVED] }
            primaryMedia: { $ref: '#/components/schemas/MediaSummary' }
            author: { $ref: '#/components/schemas/AuthorSummary' }
            isFavorited: { type: boolean }
            categories:
              type: array
              items: { $ref: '#/components/schemas/CategoryResponse' }
            tags:
              type: array
              items: { $ref: '#/components/schemas/TagResponse' }
            updatedAt: { type: string, format: date-time }

    # -----------------------------------------------------------------
    # Bulk delete request / response. Mirrors the media bulk-delete
    # shapes so the frontend can share UI + service helpers.
    # -----------------------------------------------------------------
    BulkDeleteCardsRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 50
          uniqueItems: true
          items: { type: string, format: uuid }

    BulkDeleteCardsResponse:
      type: object
      properties:
        deleted:
          type: integer
          description: Number of cards successfully deleted.
        failed:
          type: integer
          description: Number of cards that could not be deleted (permission, not found, etc.).
        errors:
          type: array
          description: One entry per failed card, in the order the failures occurred.
          items:
            type: object
            properties:
              id:    { type: string, format: uuid }
              error: { type: string }
        deletedIds:
          type: array
          description: IDs that were successfully deleted. Useful for client-side pruning.
          items: { type: string, format: uuid }

    CreateCategoryRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        description:
          type: string
          maxLength: 500
        parentId:
          type: string
          format: uuid

    CategoryResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        parentId: { type: string, format: uuid }
        subCategories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'

    CreateTagRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        slug:
          type: string
          minLength: 1
          maxLength: 50
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'

    TagResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }

    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
        validationErrors:
          type: object
          additionalProperties:
            type: string